ARM64:Defconfig:rolex: Disable modules support
Modules are a security vulnerability in Android and have been disabled
in Nexus devices for a long time. Though not restricted by Google for
known reasons, we should still try our best to reduce the number of
modules. As we no longer have any more modules, let's disable module
support.

Change-Id: I91cd45da48194ac39932ac0fa1c474749603fe44
Signed-off-by: Thagoo <lohitgowda56@gmail.com>